+++ database/cassandra.sql
@@ -1,12 +1,10 @@
 CREATE TABLE Reviews (
   id int uuid increment, -- check uuid
   restaurant_id int, 
-  user_id int,
   user text,
   location text,
   vip Boolean,
   total_reviews int,
-  username text,
   review text, 
   overall int,
   food int,
@@ -16,6 +14,6 @@ CREATE TABLE Reviews (
   noise text,
   would_recommend Boolean,
   date DATE, 
-  PRIMARY KEY (restaurant_id, user_id)
+  PRIMARY KEY (restaurant_id)
 );
